Extract scene-settings glRef syncs into headless components
Second, larger React perf step (fix #3 from the re-render analysis): carve the
scene-settings -> glRef sync effects out of the 770-line MoorhenWebMG into
dedicated null-rendering child components, so a change to any of these settings
re-renders only its small sync component instead of the whole parent.

- glRefSync/EdgeDetectSync.tsx: the 5 edge-detect effects.
- glRefSync/SceneRenderSettingsSync.tsx: the ~22 scene-render effects (SSAO,
  shadows, env-B-occ, stereo / multi-view / three-way, spin, depth blur,
  perspective projection, outlines, clip-cap, shadow-depth debug).

Each effect is copied verbatim - same guards, same setter-or-field-assignment,
same dependency arrays - so behaviour is identical; only the component the
useSelector subscription lives in has moved. React attaches all refs before
running any effect, so the child effects still see a populated glRef.current
regardless of sibling order.

MoorhenWebMG: 770 -> 575 lines, 77 -> 50 useSelectors, 58 -> 33 useEffects.
The remaining selectors are genuinely parent-coupled (props passed to MGWebGL,
viewport dims used in JSX, glRef/interaction state feeding callbacks).

Verified in-app: SSAO (toggle + radius), depth blur (toggle + sliders),
three-way view, and edge detection all still respond correctly, driven from the
extracted sync components.
